In readNextCommand() adds ';' to the "if" endline statement, if character is ';', increments processedSize and CurrentPosition until '/n' or '/r' is found. (effectively ignoring any comments received.)
If a comment is found, the command size is checked before marking
processing as finished, as full line comments would result in a command
size of 0, which would set HOST_STOP_STREAM as the next step.
With this check, it continues on to process the next line as though the
comment line wasn't there.